Use bundled version in normal install process
